unsigned int -> size_t in pointer casts
--HG--
extra : convert_revision : svn%3A7d7b52a7-431a-0410-ad82-b695db389479/trunk%40452

+++ libbibutils/newstr.c
@@ -192,9 +192,9 @@ newstr_segcat( newstr *string, char *startat, char *endat )
 	char *p, *q;
 
 	if (string==NULL || startat==NULL || endat==NULL ) return;
-	if ( (unsigned int) startat > (unsigned int) endat ) return;
+	if ( (size_t) startat > (size_t) endat ) return;
 
-	seglength=(unsigned int) endat - (unsigned int) startat;
+	seglength=(size_t) endat - (size_t) startat;
 	if (string->data==NULL || string->dim==0)
 		newstr_initalloc(string, seglength+1);
 	else {
